pyproject.toml
[tool.poetry]
name = "work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
albumentations = "~1.3.0"
torchvision = "~0.15"
segmentation-models-pytorch = "~0.3.2"
jupyterlab = "~3.6.1"
pandas = "~1.5.3"
tqdm = "~4.65.0"
matplotlib = "~3.7.1"
ipywidgets = "8.0.4"
ttach = "^0.0.3"


[tool.poetry.group.dev.dependencies]
black = "~23.1.0"
flake8 = "~6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
